Update description of the Localization pipeline
When working with GitLocalize, translators are only able to create Review Requests. Review Requests can then be converted to Pull Requests (sent to the repo) by the team members with the respective roles on Gitlocalize (Admin, Language Moderator, Owner). 
This PR updates the translation section in the contributing guide to reflect that.